# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus];

export const ContactStatus = {
  new_lead: "new_lead",
  contacted: "contacted",
  booked: "booked",
  completed: "completed",
  lost: "lost",
} as const;

export type ContactSource = (typeof ContactSource)[keyof typeof ContactSource];

export const ContactSource = {
  website: "website",
  google_ads: "google_ads",
  facebook: "facebook",
  referral: "referral",
  other: "other",
} as const;

export interface Contact {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  status: ContactStatus;
  source: ContactSource;
  /** @nullable */
  serviceRequested?: string | null;
  /** @nullable */
  notes?: string | null;
  tags: string[];
  totalRevenue: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateContactBodyStatus =
  (typeof CreateContactBodyStatus)[keyof typeof CreateContactBodyStatus];

export const CreateContactBodyStatus = {
  new_lead: "new_lead",
  contacted: "contacted",
  booked: "booked",
  completed: "completed",
  lost: "lost",
} as const;

export type CreateContactBodySource =
  (typeof CreateContactBodySource)[keyof typeof CreateContactBodySource];

export const CreateContactBodySource = {
  website: "website",
  google_ads: "google_ads",
  facebook: "facebook",
  referral: "referral",
  other: "other",
} as const;

export interface CreateContactBody {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  status?: CreateContactBodyStatus;
  source?: CreateContactBodySource;
  /** @nullable */
  serviceRequested?: string | null;
  /** @nullable */
  notes?: string | null;
  tags?: string[];
}

export type UpdateContactBodySource =
  (typeof UpdateContactBodySource)[keyof typeof UpdateContactBodySource];

export const UpdateContactBodySource = {
  website: "website",
  google_ads: "google_ads",
  facebook: "facebook",
  referral: "referral",
  other: "other",
} as const;

export interface UpdateContactBody {
  name?: string;
  phone?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  notes?: string | null;
  source?: UpdateContactBodySource;
  /** @nullable */
  serviceRequested?: string | null;
  tags?: string[];
}

export type UpdateStatusBodyStatus =
  (typeof UpdateStatusBodyStatus)[keyof typeof UpdateStatusBodyStatus];

export const UpdateStatusBodyStatus = {
  new_lead: "new_lead",
  contacted: "contacted",
  booked: "booked",
  completed: "completed",
  lost: "lost",
} as const;

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}

export interface Job {
  id: number;
  contactId: number;
  serviceType: string;
  price: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateJobBody {
  serviceType: string;
  price: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateJobBody {
  serviceType?: string;
  price?: number;
  /** @nullable */
  date?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Activity {
  id: number;
  contactId: number;
  /** @nullable */
  contactName?: string | null;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export type FollowUpStatus =
  (typeof FollowUpStatus)[keyof typeof FollowUpStatus];

export const FollowUpStatus = {
  pending: "pending",
  completed: "completed",
} as const;

export interface FollowUp {
  id: number;
  contactId: number;
  /** @nullable */
  contactName?: string | null;
  dueDate: string;
  /** @nullable */
  note?: string | null;
  status: FollowUpStatus;
  createdAt: string;
}

export interface CreateFollowUpBody {
  dueDate: string;
  /** @nullable */
  note?: string | null;
}

export type UpdateFollowUpBodyStatus =
  (typeof UpdateFollowUpBodyStatus)[keyof typeof UpdateFollowUpBodyStatus];

export const UpdateFollowUpBodyStatus = {
  pending: "pending",
  completed: "completed",
} as const;

export interface UpdateFollowUpBody {
  dueDate?: string;
  /** @nullable */
  note?: string | null;
  status?: UpdateFollowUpBodyStatus;
}

export interface DashboardSummary {
  totalLeads: number;
  bookedJobs: number;
  completedJobs: number;
  lostLeads: number;
  conversionRate: number;
  totalRevenue: number;
  overdueFollowUps: number;
}

export interface PipelineStage {
  status: string;
  count: number;
  revenue: number;
}

export interface SourceCount {
  source: string;
  count: number;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  businessName: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  businessName: string;
  createdAt?: string;
}

export type SubmitLeadBodySource =
  (typeof SubmitLeadBodySource)[keyof typeof SubmitLeadBodySource];

export const SubmitLeadBodySource = {
  website: "website",
  google_ads: "google_ads",
  facebook: "facebook",
  referral: "referral",
  other: "other",
} as const;

export interface SubmitLeadBody {
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  serviceRequested?: string | null;
  /** @nullable */
  notes?: string | null;
  source?: SubmitLeadBodySource;
}

export interface LeadSubmitResponse {
  success: boolean;
  message: string;
  contactId: number;
}

export type Logout200 = {
  success?: boolean;
};

export type ListContactsParams = {
  status?: ListContactsStatus;
  source?: ListContactsSource;
  search?: string;
  tag?: string;
};

export type ListContactsStatus =
  (typeof ListContactsStatus)[keyof typeof ListContactsStatus];

export const ListContactsStatus = {
  new_lead: "new_lead",
  contacted: "contacted",
  booked: "booked",
  completed: "completed",
  lost: "lost",
} as const;

export type ListContactsSource =
  (typeof ListContactsSource)[keyof typeof ListContactsSource];

export const ListContactsSource = {
  website: "website",
  google_ads: "google_ads",
  facebook: "facebook",
  referral: "referral",
  other: "other",
} as const;

export type ListFollowUpsParams = {
  status?: ListFollowUpsStatus;
  overdue?: boolean;
};

export type ListFollowUpsStatus =
  (typeof ListFollowUpsStatus)[keyof typeof ListFollowUpsStatus];

export const ListFollowUpsStatus = {
  pending: "pending",
  completed: "completed",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};