# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Email already exists
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new_lead, contacted, booked, completed, lost]
        - name: source
          in: query
          schema:
            type: string
            enum: [website, google_ads, facebook, referral, other]
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a new contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a single contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /contacts/{id}/status:
    patch:
      operationId: updateContactStatus
      tags: [contacts]
      summary: Change pipeline status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
  /followups:
    get:
      operationId: listFollowUps
      tags: [followups]
      summary: List all follow-ups
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed]
        - name: overdue
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of follow-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FollowUp"
  /contacts/{contactId}/followups:
    post:
      operationId: createFollowUp
      tags: [followups]
      summary: Create a follow-up reminder
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFollowUpBody"
      responses:
        "201":
          description: Created follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
  /followups/{id}:
    patch:
      operationId: updateFollowUp
      tags: [followups]
      summary: Update a follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFollowUpBody"
      responses:
        "200":
          description: Updated follow-up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowUp"
    delete:
      operationId: deleteFollowUp
      tags: [followups]
      summary: Delete a follow-up
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [new_lead, contacted, booked, completed, lost]
        source:
          type: string
          enum: [website, google_ads, facebook, referral, other]
        serviceRequested:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        totalRevenue:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - status
        - source
        - tags
        - totalRevenue
        - createdAt
        - updatedAt
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [new_lead, contacted, booked, completed, lost]
        source:
          type: string
          enum: [website, google_ads, facebook, referral, other]
        serviceRequested:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - name
        - phone
    UpdateContactBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        source:
          type: string
          enum: [website, google_ads, facebook, referral, other]
        serviceRequested:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [new_lead, contacted, booked, completed, lost]
      required:
        - status
...
    FollowUp:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
        contactName:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactId
        - dueDate
        - status
        - createdAt
    CreateFollowUpBody:
      type: object
      properties:
        dueDate:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
      required:
        - dueDate
    UpdateFollowUpBody:
      type: object
      properties:
        dueDate:
          type: string
          format: date-time
        note:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed]
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        businessName:
          type: string
      required:
        - email
        - password
        - businessName
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        businessName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - businessName
...