# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  email: string;
  businessName: string;
  createdAt?: Date;
}